v1.1.0 - Throttle, Size Limit, More Emojis
Implemented a throttle mechanism for data saving (chrome.storage.sync.set) to comply with the API request limits.
This includes both limits - 120 writes/min and 1800 writes/h.
Currently, the throttle delay is set to 2 seconds.

The data is saved immediately when closing the popup window to prevent bugs.

Added a popup to notify users when the note size exceeds the Chrome storage sync limit (in bytes).
The user will not be able to type or paste any text when it would exceed the limit.

Added more emojis.

Made some proactive improvements to reduce the likelihood of future bugs.